+++ package.json
@@ -25,16 +25,17 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@actions/core": "^1.10.0",
-    "@actions/github": "^5.1.1",
-    "@octokit/rest": "^20.0.1",
-    "@types/semver": "^7.5.0",
+   
     "compare-versions": "^6.1.0",
     "fs": "^0.0.1-security",
     "semver": "^7.5.4",
     "semver-compare": "^1.0.0"
   },
   "devDependencies": {
+    "@actions/core": "^1.10.0",
+    "@actions/github": "^5.1.1",
+    "@octokit/rest": "^20.0.1",
+    "@types/semver": "^7.5.0",
     "@types/compare-versions": "^3.3.0",
     "@types/node": "^18.17.4",
     "@typescript-eslint/parser": "^5.59.2",
