+++ src/main.ts
@@ -3,8 +3,6 @@ import {wait} from './wait'
 import {checkVersion} from './check-version'
 import {context, getOctokit} from '@actions/github'
 import * as semver from 'semver'
-import compareVersions from 'compare-versions'
-// const semverGt = require('semver/functions/gt');
 
 type GithubContext = typeof context
 
@@ -25,55 +23,55 @@ async function run(): Promise<void> {
 
     core.setOutput('time', new Date().toTimeString())
 
-    core.debug('location')
-    const myRepoURL = getRepoURL(context)
-    console.log(`This repo's URL is: ${myRepoURL}`)
+    // core.debug('location')
+    // const myRepoURL = getRepoURL(context)
+    // console.log(`This repo's URL is: ${myRepoURL}`)
 
     //processing tags
-    getTags().then(tags => {
-      if (tags) {
-        // filter out tags that don't look like releases
-        const sortedTaggedVersions = tags
-          .filter(t => t.name.match(/\d+.\d+.\d+/))
-          .sort((a, b) => semver.compare(a.name, b.name))
-        // const sortedTaggedVersions = taggedVersions.sort(semver.compare)
+    // getTags().then(tags => {
+    //   if (tags) {
+    //     // filter out tags that don't look like releases
+    //     const sortedTaggedVersions = tags
+    //       .filter(t => t.name.match(/\d+.\d+.\d+/))
+    //       .sort((a, b) => semver.compare(a.name, b.name))
+    //     // const sortedTaggedVersions = taggedVersions.sort(semver.compare)
 
-        sortedTaggedVersions.forEach(element => {
-          console.log(`
-        Your tag: \n
-        ${JSON.stringify(element, undefined, 2)}`)
-        })
-        //newest tag
-        console.log(
-          `Newest Tag: ${
-            sortedTaggedVersions[sortedTaggedVersions.length - 1].name
-          }`
-        )
-        console.log(`Oldest Tag: ${sortedTaggedVersions[0].name}`)
-      }
-    })
+    //     sortedTaggedVersions.forEach(element => {
+    //       console.log(`
+    //     Your tag: \n
+    //     ${JSON.stringify(element, undefined, 2)}`)
+    //     })
+    //     //newest tag
+    //     console.log(
+    //       `Newest Tag: ${
+    //         sortedTaggedVersions[sortedTaggedVersions.length - 1].name
+    //       }`
+    //     )
+    //     console.log(`Oldest Tag: ${sortedTaggedVersions[0].name}`)
+    //   }
+    // })
 
-    await checkVersion(location)
+    await checkVersion(location, ghToken)
   } catch (error) {
     if (error instanceof Error) core.setFailed(error.message)
   }
 }
 
-function getRepoURL({repo, serverUrl}: GithubContext): string {
-  return `${serverUrl}/${repo.owner}/${repo.repo}`
-}
-async function getTags() {
-  if (ghToken) {
-    const octokit = getOctokit(ghToken)
+// function getRepoURL({repo, serverUrl}: GithubContext): string {
+//   return `${serverUrl}/${repo.owner}/${repo.repo}`
+// }
+// async function getTags() {
+//   if (ghToken) {
+//     const octokit = getOctokit(ghToken)
 
-    const result = await octokit.rest.repos.listTags({
-      repo: context.repo.repo,
-      owner: context.repo.owner
-    })
-    // console.log(result)
+//     const result = await octokit.rest.repos.listTags({
+//       repo: context.repo.repo,
+//       owner: context.repo.owner
+//     })
+//     // console.log(result)
 
-    return result.data || []
-  }
-}
+//     return result.data || []
+//   }
+// }
 
 run()
